pyproject.toml
[tool.poetry]
name = "ssd_doghouse"
version = "0.1.0"
description = ""
authors = ["MatteoPuccio <matteopuccio2000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
coreapi = "^2.3.3"
PyYAML = "^6.0.1"
pytest-django = "^4.7.0"
mixer = "^7.2.2"
coverage = "^7.3.2"
django-extensions = "^3.2.3"
dj-rest-auth = "^5.0.2"
django-allauth = "^0.59.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
